requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-cors-headers==3.5.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
gunicorn==19.7.1
httplib2==0.18.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
paytm-pg==1.0.9
paytmchecksum==1.7.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.8
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
razorpay==1.2.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
